requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
pillow==12.1.1
numpy==2.3.4
albumentations==2.0.8
torch==2.10.0
torchvision==0.25.0
torchmetrics==1.8.2
faster-coco-eval==1.7.0
